+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix `AsyncStream.skip_while()`
+
 ## [0.0.4](https://github.com/lirsacc/sloths-py/releases/tag/0.0.4) - 2025-05-14
 
 - Add `Stream.range()`